// frontend/src/interfaces/iFormStateValidationTask.ts
import type { TFieldState } from "../types/typeStateFields";

export interface iFormStateValidationTask {
  // PASO 4(EN PROFESIONAL) BASICO(CLIENTE)
  fullName: TFieldState;
  userName: TFieldState;
  email: TFieldState;
  cityName: TFieldState;
  password: TFieldState;
  confirmPassword: TFieldState;

  // PASO 1 CATEGORIA
  category: TFieldState;
  service: TFieldState;
  workArea: TFieldState;
  day: TFieldState;
  hour: TFieldState;

  // PASO 2 PERFIL
  description: TFieldState;
  imageProfileData: TFieldState;
  imageExperienceData: TFieldState;

  // PASO 3 (PRESUPUESTO SI INCLUYE SOLO VALIDAR EL CAMPO DEL MONTO)
  amountBudge: TFieldState;
  budgeSelected: TFieldState;
  reinsert: TFieldState;
}